+++ aten/src/ATen/native/vulkan/glsl/vulkan_image_to_nchw.glsl
@@ -1,7 +1,7 @@
 #version 450 core
 layout(std430) buffer;
 layout(std430) uniform;
-layout(set = 0, binding = 0) uniform mediump sampler3D uInput;
+layout(set = 0, binding = 0) uniform highp sampler3D uInput;
 layout(set = 0, binding = 1) writeonly buffer destBuffer {
   float data[];
 }
